requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
alembic
python-multipart
pytest
httpx
